refactored index balance new by moving logic funcs to helpers.ts
